fix(web-shell): package ui primitive validator
