+++ MANIFEST.in
@@ -5,7 +5,7 @@ recursive-include factory_app/refinement_harness *
 recursive-include factory_app/workflows *
 recursive-include mozaiksai/core/usage/catalogs *.json
 include pricing/catalogs/usage-pricing.generated.json
-recursive-include web_shell *.js *.jsx *.css *.html *.json *.md
+recursive-include web_shell *.js *.jsx *.cjs *.css *.html *.json *.md
 include web_shell/.env.example
 recursive-include chat-ui/src *
 include chat-ui/package.json
