+++ tests/test_release_packaging_contract.py
@@ -1,5 +1,6 @@
 from __future__ import annotations
 
+import json
 from pathlib import Path
 
 ROOT = Path(__file__).resolve().parents[1]
@@ -23,9 +24,18 @@ def test_setup_excludes_generated_frontend_dependency_trees() -> None:
 def test_manifest_includes_packaged_frontend_sources() -> None:
     manifest = (ROOT / "MANIFEST.in").read_text(encoding="utf-8")
 
-    assert "recursive-include web_shell *.js *.jsx *.css *.html *.json *.md" in manifest
+    assert "recursive-include web_shell *.js *.jsx *.cjs *.css *.html *.json *.md" in manifest
     assert "recursive-include chat-ui/src *" in manifest
     assert "include chat-ui/package.json" in manifest
+    assert (ROOT / "web_shell" / "scripts" / "validate-ui-primitive-usage.cjs").exists()
+
+
+def test_web_shell_ui_primitive_script_is_package_relative() -> None:
+    package_json = json.loads((ROOT / "web_shell" / "package.json").read_text(encoding="utf-8"))
+
+    assert package_json["scripts"]["test:ui-primitives"] == (
+        "node ./scripts/validate-ui-primitive-usage.cjs"
+    )
 
 
 def test_manifest_includes_factory_defaults_used_by_app_overlays() -> None:
